[#156913774] implement AddPaymentMethodScreen to be used in transaction
