+++ ts/navigation/AppNavigator.ts
@@ -1,5 +1,4 @@
 import { SwitchNavigator } from "react-navigation";
-
 import IngressScreen from "../screens/IngressScreen";
 import AuthenticationNavigator from "./AuthenticationNavigator";
 import MainNavigator from "./MainNavigator";
