+++ ts/navigation/routes.ts
@@ -32,7 +32,7 @@ const ROUTES = {
   WALLET_ADD_CARD: "WALLET_ADD_CARD",
   WALLET_CONFIRM_TO_PROCEED: "WALLET_CONFIRM_TO_PROCEED",
   WALLET_PAY_WITH: "WALLET_PAY_WITH",
-
+  
   // Pin
   PIN_LOGIN: "PIN_LOGIN",
